Adicionadas funcionalidades do modo preview para o APK de preview.

+++ src/Features/User-CRUD/Cadastro/DefinePhotoForm.tsx
@@ -1,9 +1,11 @@
 import { Text, View } from "react-native";
 import { router } from "expo-router";
 
+import { env } from "Config/env";
+
+import { useAuth } from "@Auth/Contexts/useAuth";
 import { useShowLoadingScreen } from "Contexts/useShowLoadingScreen";
 import { useInternetConnection } from "Contexts/useInternetConnection";
-import { useAuth } from "@Auth/Contexts/useAuth";
 import { useSelectImage } from "@Auth/Hooks/useSelectImage";
 
 import Button from "@Components/Button";
@@ -14,7 +16,7 @@ import { ImageContainer } from "../../Auth/Components/ImageContainer";
 const { container, subtitle } = DefinePhotoFormStyles;
 
 export default function DefinePhotoForm() {
-  const { temporaryImageToken } = useAuth();
+  const { temporaryImageToken, login, user } = useAuth();
   const { handleImageSending, handleImageSubmit, imageURI } = useSelectImage(
     "upload-user-img",
     "POST",
@@ -23,12 +25,21 @@ export default function DefinePhotoForm() {
   const { getIsConnected } = useInternetConnection();
 
   async function handleSubmit() {
+    if (env.buildProfile === "preview") {
+      await login({ ...user, user_img: imageURI });
+      router.push("/Content");
+      return;
+    }
     if (!getIsConnected()) return;
     setShowLoadingScreen(true);
+
     const userToken = await temporaryImageToken.get();
+
     await handleImageSubmit(userToken);
     await temporaryImageToken.delete();
+
     router.replace("/Login");
+
     setShowLoadingScreen(false);
   }
